Crate cuckoofilter

Cuckoo filter probabilistic data structure for membership testing and cardinality counting.

Usage

This crate is on crates.io and can be used by adding cuckoofilter to the dependencies in your project’s Cargo.toml.

[dependencies]
cuckoofilter = "0.3"

And this in your crate root:

extern crate cuckoofilter;

Structs

CuckooFilter

A cuckoo filter class exposes a Bloomier filter interface, providing methods of add, delete, contains.

ExportedCuckooFilter

A minimal representation of the CuckooFilter which can be transfered or stored, then recovered at a later stage.

Enums

CuckooError

Constants

DEFAULT_CAPACITY

The default number of buckets.

MAX_REBUCKET

If insertion fails, we will retry this many times.